requirements.txt
# GPU版本PaddlePaddle
paddlepaddle-gpu>=3.2.0

# OCR
paddleocr>=2.9.0

# 网络分析
scapy>=2.5.0
pyshark>=0.5.0

# 图像处理
opencv-python>=4.9.0
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 系统监控
psutil>=5.9.0

# 进度条
tqdm>=4.66.0

# 正则表达式
regex>=2024.0.0

# HTTP请求
requests>=2.31.0

# 可视化
matplotlib>=3.7.0

# GPU监控
pynvml>=12.0.0
